Remove `Sets.complementOf` from `guava-gwt`.
Previously, it was present but failed at runtime under J2CL. And soon, J2CL may drop support for the reflective enum-related APIs that it depends on.

`guava-gwt` is used by both GWT and J2CL users, so we have to keep our code buildable against both.

If this CL causes trouble for you as a GWT user, let us know, and we can investigate options further. In the worst case, you can copy the implementation as needed, since `complementOf` is a static method.

RELNOTES=`collect`: Removed `Sets.complementOf` from `guava-gwt` to prepare for the removal of reflective enum-related APIs from J2CL. If this causes you problems as a GWT user, let us know.
PiperOrigin-RevId: 535589407