Automatically bump deps, and manually bump `error_prone_core`.
Then I had to update the version numbers in `build.gradle.kts`. In the future, we could consider:
- having the Gradle script read them from the `pom.xml` (as it already does for the _Guava_ version, though the need to access `pom.xml` is awkward)
- having our test process read them from the `pom.xml` with some Maven command run from the shell and then passing the result to Gradle as some kind of command-line property (which we could presumably also use for passing the _Guava_ version)
- having our _upgrade_ process read them as above but then update the Gradle script text (as I'm doing in this CL, just automatic)
- modifying the Gradle test to not care about the versions of those libraries (but to continue caring about the version of `guava` and `listenablefuture`)

And it looks like `versions-maven-plugin` bumped a few plugins, too? Maybe we could get it to bump _all_ plugins if we declared a `property` for each of them (as long as the `plugin` appears in the same `pom.xml` as the `property`, given our discovery about properties in cl/552479838)? Possible future work...

RELNOTES=n/a
PiperOrigin-RevId: 554548816