Give `ContiguousSetTest` its own `DiscreteDomain` implementation instead of pulling one from `RangeTest`.
By having the tests isolated from one another, we make it easier to run them in our forthcoming J2CL setup.

Rather than copy the implementation exactly, I changed it to make it throw if any unexpected methods are called. That's not really necessary, but it seemed like an extra layer of testing that the `IllegalArgumentException` is happening for the reason we expect. And the important thing is that it made me feel like I wasn't just copying and pasting code :)

PiperOrigin-RevId: 570503073