Prepare to make `ImmutableList.toImmutableList` available in guava-android (but don't do so yet).
This CL also includes updates to avoid problems with reflection when Java 8+ APIs are not present. That means adding a `serialVersionUID` (so that Java/Android reflection doesn't try to compute one with reflection). (Compare cl/550872250.)

This CL is further progress toward https://github.com/google/guava/issues/6567

RELNOTES=n/a
PiperOrigin-RevId: 576885237