Document and test how `ImmutableMap.toImmutableMap` behaves when the `mergeFunction` returns `null`.
(The test is Google-internal for now because we're in the process of reshuffling our `Collector` tests to make them run under Android as part of https://github.com/google/guava/issues/6567. Also, we skip the test under GWT+J2CL because of a bug in their implementation of `Collectors.toMap`.)

This addresses the main part of https://github.com/google/guava/issues/6824, but I'm keeping the bug open as a prompt to recognize our nullness annotations in the future.

RELNOTES=n/a
PiperOrigin-RevId: 580661189