Simplify and micro-optimize `Iterators.forArray`.
- Size before: 32 bytes
- Size after: 24 bytes

(There remains an unnecessary `AbstractIndexedListIterator.size` field (which is always set to the array length). But Java object sizes in RAM are aligned to whole words (at least in typical environments), it's currently the second of two 4-byte fields that make up the final such word. So even if we eliminate it, the object uses the whole word. Plus, to eliminate that field, we'd probably need to duplicate some code.)

I noticed this while debugging a problem with cl/590212390. Naturally, it's not a big deal.

The simplification is nice, at least. It includes renaming one of the `forArray` overloads so that `forArray(array, integer)` can mean only one thing.

RELNOTES=n/a
PiperOrigin-RevId: 591325737