Add `Duration` overload for `Suppliers.memoizeWithExpiration`.
This CL contains part but not all of https://github.com/google/guava/pull/3691

Fixes https://github.com/google/guava/issues/3169

We'd considered this previously in cl/197933201 but gotten stuck on the question of whether to accept and/or return the `java.util.function` type. My claim is that we want an overload that is near-identical to the old one for ease of migration. Additionally, if we want to support `java.util.function.Supplier` in the future, then we're going to have to add an overload of the existing `memoize` method, and which point it would probably look _more_ weird if we had 2 overloads for `memoize` (and maybe `synchronizedSupplier`) but only a single `memoizeWithExpiration` method.

RELNOTES=`base`: Added a `Duration` overload for `Suppliers.memoizeWithExpiration`.
PiperOrigin-RevId: 597280125