Fix testing-j2objc when built with no-wrapper-methods
This target previously broke when built with --j2objc_translation_flags=--no-wrapper-methods. The root cause appears to be special handling of methods that begin with "new". This CL changes the name of the private "newMatchResult" method to "createMatchResult" which eliminates the issue.

RELNOTES=Allows building testing package with --j2objc_translation_flags=--no-wrapper-methods
PiperOrigin-RevId: 608726803